增加条件编译，修复windows下gcc依赖问题
